Gemini Clone with voice/requirements.txt
Streamlit
genai
google-generativeai
pyttsx3
SpeechRecognition